Allowed transactions and revisions to include (limited) schema upgrades.

+++ iModelCore/BeSQLite/BeSQLite.cpp
@@ -2000,11 +2000,29 @@ DbResult Db::DeleteBriefcaseLocalValue(Utf8CP name)
     }
 
 /*---------------------------------------------------------------------------------**//**
+* @bsimethod                                  Ramanujam.Raman                   11/16
++---------------+---------------+---------------+---------------+---------------+------*/
+bool isTempTableOrIndex(Utf8CP tableName)
+    {
+    return 0 == BeStringUtilities::Strnicmp(tableName, "temp.", 5);
+    }
+
+/*---------------------------------------------------------------------------------**//**
 * @bsimethod                                    Keith.Bentley                   12/10
 +---------------+---------------+---------------+---------------+---------------+------*/
 DbResult Db::CreateTable(Utf8CP tableName, Utf8CP ddl) const
     {
-    return ExecuteSql(SqlPrintfString("CREATE TABLE %s (%s)", tableName, ddl));
+    SqlPrintfString sql("CREATE TABLE %s (%s)", tableName, ddl);
+    
+    DbResult result = ExecuteSql(sql);
+    if (result != BE_SQLITE_OK)
+        return result;
+
+    ChangeTracker* tracker = m_dbFile->m_tracker.get();
+    if (tracker && !isTempTableOrIndex(tableName))
+        result = tracker->RecordSchemaChange(sql.GetUtf8CP());
+
+    return result;
     }
 
 /*---------------------------------------------------------------------------------**//**
@@ -2012,9 +2030,55 @@ DbResult Db::CreateTable(Utf8CP tableName, Utf8CP ddl) const
 +---------------+---------------+---------------+---------------+---------------+------*/
 DbResult Db::DropTable(Utf8CP tableName) const
     {
+    ChangeTracker* tracker = m_dbFile->m_tracker.get();
+    if (tracker && tracker->IsTracking() && !isTempTableOrIndex(tableName))
+        {
+        BeAssert(false && "Cannot make arbitrary schema changes when changes are being tracked");
+        return BE_SQLITE_ERROR;
+        }
+
     DbResult rc = TryExecuteSql(SqlPrintfString("DROP TABLE %s", tableName));
     BeAssert(rc == BE_SQLITE_OK);
     return rc;
+    }
+
+/*---------------------------------------------------------------------------------**//**
+* @bsimethod                                  Ramanujam.Raman                   01/17
++---------------+---------------+---------------+---------------+---------------+------*/
+DbResult Db::AddColumnToTable(Utf8CP tableName, Utf8CP columnName, Utf8CP columnDetails)
+    {
+    SqlPrintfString sql("ALTER TABLE %s ADD COLUMN %s %s", tableName, columnName, columnDetails);
+
+    DbResult result = ExecuteSql(sql);
+    if (result != BE_SQLITE_OK)
+        return result;
+
+    ChangeTracker* tracker = m_dbFile->m_tracker.get();
+    if (tracker && !isTempTableOrIndex(tableName))
+        result = tracker->RecordSchemaChange(sql.GetUtf8CP());
+
+    return result;
+    }
+
+/*---------------------------------------------------------------------------------**//**
+* @bsimethod                                  Ramanujam.Raman                   01/17
++---------------+---------------+---------------+---------------+---------------+------*/
+DbResult Db::CreateIndex(Utf8CP indexName, Utf8CP tableName, bool isUnique, Utf8CP columnName1, Utf8CP columnName2 /*=nullptr*/)
+    {
+    BeAssert(!Utf8String::IsNullOrEmpty(indexName) && !Utf8String::IsNullOrEmpty(tableName) && !Utf8String::IsNullOrEmpty(columnName1));
+    
+    // e.g.,  "CREATE UNIQUE INDEX ix_ec_Class_SchemaId_Name ON ec_Class(SchemaId, Name);"
+    SqlPrintfString sql("CREATE %s INDEX %s ON %s(%s%s%s)", isUnique ? "UNIQUE" : "", indexName, tableName, columnName1, Utf8String::IsNullOrEmpty(columnName2) ? "" : ",", columnName2 ? columnName2 : "");
+
+    DbResult result = ExecuteSql(sql);
+    if (result != BE_SQLITE_OK)
+        return result;
+
+    ChangeTracker* tracker = m_dbFile->m_tracker.get();
+    if (tracker && !isTempTableOrIndex(indexName))
+        result = tracker->RecordSchemaChange(sql.GetUtf8CP());
+
+    return result;
     }
 
 //---------------------------------------------------------------------------------------
@@ -2073,9 +2137,16 @@ bool Db::GetColumns(bvector<Utf8String>& columns, Utf8CP tableName) const
 /*---------------------------------------------------------------------------------**//**
 * @bsimethod                                                    Affan.Khan      10/2013
 +---------------+---------------+---------------+---------------+---------------+------*/
-bool Db::RenameTable(Utf8CP tableName, Utf8CP newTableName)
+DbResult Db::RenameTable(Utf8CP tableName, Utf8CP newTableName)
     {
-    return BE_SQLITE_OK == ExecuteSql(SqlPrintfString("ALTER TABLE %s RENAME TO %s", tableName, newTableName));
+    ChangeTracker* tracker = m_dbFile->m_tracker.get();
+    if (tracker && tracker->IsTracking() && !isTempTableOrIndex(tableName))
+        {
+        BeAssert(false && "Cannot make arbitrary schema changes when changes are being tracked");
+        return BE_SQLITE_ERROR;
+        }
+
+    return ExecuteSql(SqlPrintfString("ALTER TABLE %s RENAME TO %s", tableName, newTableName));
     }
 
 /*---------------------------------------------------------------------------------**//**
